migrates remaining components
