+++ apps/globetrotter/src/app/app-routing.module.ts
@@ -1,10 +1,9 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes, PreloadAllModules } from '@angular/router';
 
-import { HomeComponent, PageNotFoundComponent } from '@atocha/globetrotter-feature-shell';
+import { HomeComponent, PageNotFoundComponent, ShellComponent } from '@atocha/globetrotter-feature-shell';
 import { CountryService } from '@atocha/globetrotter-data-access';
 import { Route } from '@atocha/globetrotter-types';
-import { ShellComponent } from './components/shell/shell.component';
 
 const routes: Routes = [
   {
