+++ libs/globetrotter/feature-shell/src/index.ts
@@ -1,4 +1,6 @@
 export * from './lib/globetrotter-feature-shell.module';
 
+export * from './lib/error/error.component';
 export * from './lib/home/home.component';
 export * from './lib/page-not-found/page-not-found.component';
+export * from './lib/shell/shell.component';
