+++ libs/globetrotter/feature-shell/src/lib/error/error.component.html
@@ -1,6 +1,6 @@
 <div class="content"
   @fadeIn
-  >
+>
   <h1 class="header">
     Server Error
   </h1>
