+++ libs/globetrotter/feature-shell/src/lib/error/error.component.ts
