+++ libs/globetrotter/feature-shell/src/lib/navigation/navigation.component.html
@@ -3,7 +3,7 @@
   (@position.done)="onAnimationFinish($event)"
   >
   <a class="home"
-    routerLink="/{{home.route}}"
+    routerLink="/{{ home.route }}"
     >
     <ui-icon *ngIf="home.icon"
       class="home__icon"
