+++ libs/globetrotter/feature-shell/src/lib/navigation/navigation.component.scss
