+++ libs/globetrotter/feature-shell/src/lib/shell/shell.component.scss
